Add Slack slash command & app example.
To use, edit the config in `config.go` and deploy to App Engine with
`goapp deploy -application YOUR-PROJECT-ID app.yaml`.

We add an "Add to Slack" button and show how to handle the Oauth flow,
using the `golang.org/x/oauth2` client library.  Theoretically, this
could be done for the other examples, but the slash command is easiest
since it doesn't actually require interfacing with any external state.