Organize samples by programming language.
I think it's more likely that people will be looking for a sample in a
language they know than it is they are looking for a sample showing a
specific feature in any language.

If we get a whole bunch for a single language, we might want to break it
down further, but this seems fine for now.

Also, now that the community site is live, move the tutorial for the
slack notifications to there instead. Fewer pages to keep in sync that
way.